refactor: Menu - MenuGroup - MenuProduct 간 양방향 및 객체참조 제거
